% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_data.R
\name{mod_compare_data_ui}
\alias{mod_compare_data_ui}
\alias{mod_compare_data_server}
\title{mod_compare_data_ui and mod_compare_data_server}
\usage{
mod_compare_data_ui(id)

mod_compare_data_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
