% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_html.R
\name{mod_export_html_ui}
\alias{mod_export_html_ui}
\alias{mod_export_html_server}
\title{mod_export_html_ui and mod_export_html_server}
\usage{
mod_export_html_ui(id)

mod_export_html_server(input, output, session, html, filename)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
