% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_xlsx.R
\name{mod_export_xlsx_ui}
\alias{mod_export_xlsx_ui}
\alias{mod_export_xlsx_server}
\title{mod_export_xlsx_ui and mod_export_xlsx_server}
\usage{
mod_export_xlsx_ui(id)

mod_export_xlsx_server(input, output, session, old_dups, new_dups,
  filename)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
