% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_table}
\alias{show_table}
\title{Render DT datable with export button and scrollX enabled}
\usage{
show_table(df)
}
\arguments{
\item{df}{}
}
\value{
a DT datatable
}
\description{
Render DT datable with export button and scrollX enabled
}
